Implemented shared Drawer component Done some minor changes to form components
