+++ components/page/demo-day/FounderPendingView/components/EditProfileDrawer/EditProfileDrawer.module.scss
@@ -71,34 +71,6 @@
 
 
 // Drawer Styles
-.drawerOverlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: rgba(0, 0, 0, 0.5);
-  backdrop-filter: blur(4px);
-  -webkit-backdrop-filter: blur(4px);
-  z-index: 1000;
-  display: flex;
-  justify-content: flex-end;
-}
-
-.drawerContainer {
-  background: white;
-  width: 720px;
-  height: 100vh;
-  overflow-y: auto;
-  display: flex;
-  flex-direction: column;
-  box-shadow: -4px 0 20px rgba(0, 0, 0, 0.15);
-
-  @include media.mobile-only {
-    width: 100vw;
-  }
-}
-
 .drawerHeader {
   display: flex;
   flex-direction: column;
